팀전 랭킹 점수에 팀 강함 비율 적용
팀전은 경기마다 '팀 강함 비율' f = (진 팀 강함 합) ÷ (이긴 팀 강함 합)을
곱한다. 강한 팀으로 약팀을 이기면 f가 작아 적게 얻고, 약한 팀으로 강팀을
이기면 f가 커서 크게 얻는다(강한 팀이 지면 크게 잃고 약한 팀이 지면 적게
잃는다). 개인전(1:1)은 f=1로 예전 점수 그대로.

- repository.rank_scoring_rows: 경기별 라인업(match_id/team/member_pk/race/result)
  원본 — 팀 강함 비율을 경기마다 매기려면 head_to_head처럼 쌍으로 뭉친 값이
  아니라 경기 단위 라인업이 필요하다.
- service._apply_rank_order: 점수를 경기 단위로 합산하고 팀전엔 f를 곱한다.
  비율 때문에 소수가 되므로 최종 점수는 소수 첫째 자리에서 반올림한다.
- schemas: rank_score를 float로.

+++ src/app/domain/matches/schemas.py
@@ -187,9 +187,10 @@ class MemberStatsEntry(BaseModel):
     superior_count: int | None = Field(default=None, alias="superiorCount")
     equal_count: int | None = Field(default=None, alias="equalCount")
     inferior_count: int | None = Field(default=None, alias="inferiorCount")
-    # 랭킹 총점 — 경기마다 (이김 +2·강함 / 비김 +1·강함 / 짐 -1·약함)을 합산한 값. 카드에
-    # 이 숫자만 보여주고 세부는 상세 화면에서 본다(요청). 음수 가능. 순위 대상이 아니면 None.
-    rank_score: int | None = Field(default=None, alias="rankScore")
+    # 랭킹 총점 — 경기마다 (이김 +강함(상대) / 비김 0 / 짐 −약함(상대))을 합산하되, 팀전은
+    # 팀 강함 비율(진 팀÷이긴 팀 강함 합)을 곱한다. 카드에 이 숫자만 보여주고 세부는 상세에서
+    # 본다(요청). 팀 강함 비율 때문에 소수(첫째 자리 반올림) 가능·음수 가능. 순위 대상 아니면 None.
+    rank_score: float | None = Field(default=None, alias="rankScore")
 
 
 class MatchStatsResponse(BaseModel):
